mitosheet: add first pro feature, excel export formatting
