+++ mitosheet/mitosheet/user/__init__.py
@@ -6,7 +6,7 @@ from mitosheet.user.db import (
 )
 from mitosheet.user.utils import (
     is_running_test, is_on_kuberentes_mito, is_local_deployment,
-    should_upgrade_mitosheet
+    should_upgrade_mitosheet, is_pro
 )
 from mitosheet.user.schemas import (
     UJ_INTENDED_BEHAVIOR, UJ_CLOSED_FEEDBACK, UJ_MITOSHEET_LAST_FIVE_USAGES,
